+++ trace/google/cloud/trace/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Google LLC
+# Copyright 2017, Google LLC All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,10 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from __future__ import absolute_import
+
 from google.cloud.trace.client import Client
+from google.cloud.trace_v2 import types
+from google.cloud.trace_v2.gapic import enums
+from google.cloud.trace_v2.gapic import trace_service_client
 
 
-__all__ = ['Client', 'SCOPE']
+class TraceServiceClient(trace_service_client.TraceServiceClient):
+    __doc__ = trace_service_client.TraceServiceClient.__doc__
+    enums = enums
 
+
+__all__ = (
+    'enums',
+    'types',
+    'TraceServiceClient',
+    'Client',
+    'SCOPE',)
 
 SCOPE = Client.SCOPE
